Fix NFS "already mounted" detection
Previous fix

2f32c98a RemoteFS: don't fail in do_mount if already mounted

was made with the assumption that "already mounted" appearing
in stderr from an NFS mount means that the NFS share was already
mounted.

However, the NFS client can fail with "busy or already mounted"
in cases where it fails to mount the share as well.

This results in the Cinder NFS backup driver believing that the
NFS mount has succeeded when it didn't, and as a result, data
can be written to the local disk instead of the NFS target.

Fail if the share isn't actually mounted rather than succeeding.

Related-Bug: #1780813
Closes-Bug: #1797233
Change-Id: Iebd1afb3340fcaeb1969784966c4f9be35a28417
(cherry picked from commit 360a20c7c7b4c385255200a574edc0ca1ac8d87c)