Windows SMBFS: fix using share subdirs
When using Scale-Out File Server SMB shares, we have to use the local
share paths instead of the UNC paths if the host is part of the
cluster.

There's an os-win method that allows fetching the shared path but
it expects the share name, while we're passing subdirs as well,
for which reason it fails.

We'll have to make sure we're passing the share name only, appending
the subdir afterwards.

Closes-Bug: #1805341

Change-Id: I47d7e7afefe050cd4b25435cfcfc3e8963bd04bc