Make sure looping calls are properly mocked
The test test_connect_volume_device_not_valid in
FibreChannelConnectorTestCase was taking over 6 seconds to complete
since the FixedIntervalLoopingCall in the method under test was not
being mocked out, making the test hang until the timeout was reached.

This fixes this and any other connector test case by moving the mocking
with a test fake into the base connector test class setUp call.

Change-Id: I6b174d94b5518d6267c1fa75528a90d95a312419
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>