Fix VxFlexOs KeyError after upgrade
A long time ago the driver was changed to store volume_id in the
nova database and use that. Previously it would use the volume_name
and make a call to get the volume_id.

Both should be supported, but as the code path currently
throws a KeyError it will never reach the backwardscompatible
code.

This makes sure that the driver works even after an upgrade
from Mitaka.

Change-Id: Ic389e1a8f0f43ef410eee89202fef19d554fce90
Closes-Bug: #1648629