Revert "Fix VxFlexOs KeyError after upgrade"
This reverts commit 6b60614afaf88819ce92a2d0ed269035a3c8b261.

The Cinder ScaleIO driver has been rebranded to VxFlexOS, which is fine.
What shouldn't have happened is that the rebranding made it's way
into os-brick, resulting in a new connector protocol and mapping.
The new VxFlexOS driver in cinder should just use the existing
scaleio connector protocol, as it's not customer/user facing, and
ensures backwards and forwards compatibility.

Change-Id: Icf7063eb47368cb7b0ffb2e3f5e30903797c7b34