luks: Default to LUKS v1 when formatting volumes
cryptsetup 2.0.6 has started to default to LUKS v2 or LUKS2 when
formatting volumes.  However QEMU currently has no support for v2 and
will fail to open the volume when being used to natively decrypt.

For now we should default the original LUKS encryptor to v1 when
formatting volumes, introducing a new LUKS2 encryptor and encryption
constant in a future change.

Closes-bug: 1831994
Change-Id: I152fe10ff5a3131950b789d3fd4efa15c554ff09