Make NFS already mounted message debug level
When a share is already mounted there is currently a message logged at
info level stating this whenever mount is called. Mount is called for
every remote share when collecting stats, so this results in
unactionable and useless messages filling the log file once a minute by
default.

Since nothing needs to be done in this case, and it is not really useful
information to the administrator, this changes that logging to debug
level so it is only seen if enabling the verbosity where an admin would
want to see something like this.

Change-Id: If7dc84c50ca7c9bae878a72b330c6d5ef9a3bab1
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>