encryptors: Deprecate the CryptsetupEncryptor
This encryptor and the underlying plain cryptsetup encryption format it
provides is not widely used, inflexible (no key rotation) and as a
result no longer required in os-brick. This change starts the
deprecation process.

Additional changes will be posted to ensure the retype workflow between
this encryptor and the LUKS based encryptors is well tested ahead of any
removal.

Change-Id: Ibb560da269a2f330526af6761fa509c262e3d361