FC improve logging
This patch improves logging for the FC connector and simplifies a couple
of list and dict creations using generators.

The logging is around the modification that os-brick does of the
connection dictionary that Cinder sent.  It displays what the contents
were on entry and on exit, and it also warns if we are dropping some
targets because the Cinder driver returned inconsistent data.

Change-Id: I44bfd4939a850e8e080c4f8760244146a75f7218