rbd: Use showmapped to find the root RBD device during disconnect_volume
The original method of unmapping the higher level RBD device found under
/dev/rbd/{pool}/{volume} fails when an encryptor has been attached to
the volume locally. This is due to a symlink being created that points
to the the decrypted dm-crypt device, breaking any attempt to unmap the
RBD volume.

To avoid this we can simply find and unmap the lower level RBD device of
/dev/rbd*.

Change-Id: Id507109df80391699074773f4787f74507c4b882