iSCSI detect multipath DM with no WWN
If udev rules are slow (or don't happen) and they don't generate the
symlinks we use to detect the WWN, or if we just fail to find the WWN we
end up not detecting the iSCSI multipath even if it is present in the
system.

With this patch we no longer wait to find the WWN before trying to
locate the multipath device.

This means that as long as the multipath daemon is able to generate the
multipath we will be able to return a multipath device path to the
caller.

This backport has been tuned to properly work with Python 2 too, where
the number of calls to get_wwn_mock may be higher than 2
because the thread may not start immediately as it happens on Python 3.x.

Closes-Bug: 1881619
Change-Id: Ic48bd9ac408c56073e58168df7e74e4b949ac2f2
(cherry picked from commit 63f52be546b4bb4689e400521bea48ce25daa703)
(cherry picked from commit a04e553add6c2258efd7f044248ac5b401166189)