iSCSI detect multipath DM with no WWN
If udev rules are slow (or don't happen) and they don't generate the
symlinks we use to detect the WWN, or if we just fail to find the WWN we
end up not detecting the iSCSI multipath even if it is present in the
system.

With this patch we no longer wait to find the WWN before trying to
locate the multipath device.

This means that as long as the multipath daemon is able to generate the
multipath we will be able to return a multipath device path to the
caller.

Closes-Bug: 1881619
Change-Id: Ic48bd9ac408c56073e58168df7e74e4b949ac2f2
(cherry picked from commit 63f52be546b4bb4689e400521bea48ce25daa703)