RBD: Support non OpenStack usage
The RBD connector can get a handle to a volume when we have the config
and keyring file in the host and when all the information is provided by
the Cinder driver (including the keyring).

The code for the local attach (using krbd) does not work if all the
information is provided by the Cinder driver, which is the case of
cinderlib.

This patch ensures that the local attach/detach behaves in the same way
as when we request a handle.

For security reasons, when doing a local attach for non-OpenStack usage
the configuration file which contains the keyring will be owned by root
instead of the normal user.

Related-Bug: #1883720
Change-Id: I315f474979ea5e797e3551cdc56ef136fa7a67c0