ScaleIO: Connection info backward compatibility
When we fixed bug 1823200 in Change-ID
Iab54c515fe7be252df52b1a0503a251779805759 we made the ScaleIO connector
incompatible with the old connection properties dictionary as it only
supported the new 'config_group' and 'failed_over' parameters to get the
password.

This is a problem in any system that is upgraded and has attachments to
the array, because the connection properties of those volumes will not
contain the new fields and detaching them will result in error
"KeyError: 'config_group'".

This patch adds compatibility code to support the old connection
properties format so we can detach those volumes.

Related-Bug: #1823200
Change-Id: I6f01a178616b74ed9a86876ca46e7e46eb360518