New fix for rbd connector to work with ceph octopus
Found while testing cinder-backup with ceph on Ubuntu Focal, which
installs ceph Octopus.  Octopus apparently enforces the requirement
that the config file contain a '[global]' section for general
requirements.  The '[global]' section goes back at least to ceph
Hammer [0], so we will simply add it to the temporary ceph config
file that os-brick generates in the RBDConnector class.

[0] https://docs.ceph.com/docs/hammer/rados/configuration/mon-config-ref/

Co-authored-by: Alex Kavanagh <alex@ajkavanagh.co.uk>
Co-authored-by: Ivan Kolodyazhny <e0ne@e0ne.info>

Change-Id: I86eb31535d990291945de5d9846b1a03157ec2cf
Closes-bug: #1865754