Avoid volume extension errors caused by multipath-tools version
`multipathd reconfigure` is an asynchronous command as of
multipath-tools 0.6.1 [1][2], potentially even before that [3].

Extending in-use iSCSI or FC volumes can fail because of that
as `multipathd resize map` will output "timeout" while the
"multipathd reconfigure" operation is still in progress.

This commit will ensure that multipathd errors are handled
accordingly, retrying in case of timeouts for up to 2 minutes.

[1] https://github.com/openSUSE/multipath-tools/blob/0.6.0/multipathd/main.c#L997
[2] https://github.com/openSUSE/multipath-tools/blob/0.6.1/multipathd/main.c#L1135
[3] https://github.com/opensvc/multipath-tools/blob/b21c567961f518810a1ac3b209c8db45f6dbac2c/multipathd/cli_handlers.c#L847-L851

Change-Id: I66e866700728eee7160f48455258c3974ada55bf
Closes-Bug: #1888675