NVMe-oF: Flush on disconnect
The current NVMe-oF connector code in os-brick doesn't flush on
disconnect, as it just calls

  nvme disconnect -n <conn_nqn>

Which doesn't flush on the `nvme` code or the kernel code.

This patch adds the normal block device ioctl flush call:

    blockdev --flushbufs /dev/nvme0n1

The patch also adds and extends disconnect unit tests that were removed
when we added the new NVMe connector code (were not removed when we
added old code back).

Closes-Bug: #1903032
Change-Id: Ie6c8bdf1f3a629206da019405aee11c802a6a9bb