Failure to generate hostnqn in case missing "show-hostnqn" sub-command
Environment:
- openstack version: Xena
- kolla on ubuntu focal 20.04
the installed version of nvme is 1.9, this version doesn't have the
command "show-hostnqn".
In case of missing hostnqn file, it will fail to create hostnqn
the exit error code is 231 and is not caught by the
try-except block.

Here we try to parse the stdout to check if the
subcommand (show-hostnqn) is not supported.
if its not, we continue to try and create the hostnqn.

Change-Id: I318f167baa0ba7789f4ca2c7c12a8de5568195e0