Update requirements minima for Yoga release
Had to raise min oslo.vmware to 3.10.0 or can't run on py3.9.
So while I was in there, looked at the other requirements.
Updated most of the openstack libraries, left the third party
ones alone (see notes below).

requirements
------------
pbr>=5.5.1                -> >=5.8.0      in u-c for 3 months
oslo.concurrency>=4.4.0   -> >=4.5.0      in u-c for 4 months
oslo.context>=3.1.1       -> >=3.4.0      in u-c for 4 months
oslo.log>=4.4.0           -> >=4.6.1      in u-c for 4 months
oslo.i18n>=5.0.1          -> >=5.1.0      in u-c for 5 months
oslo.privsep>=2.4.0       -> >=2.6.2      in u-c for 6 months
oslo.serialization>=4.1.0 -> >=4.2.0      in u-c for 6 months
oslo.service>=2.5.0       -> >=2.8.0      in u-c for 1 month
oslo.utils>=4.8.0         -> >=4.12.1     in u-c for 3 weeks
os-win>=5.4.0             -> >=5.5.0      in u-c for 6 months

eventlet >=0.30.1  ->  >=0.30.1,!=0.32.0
- 0.33.0 has been in u-c for about 3 months
- 0.32.0 breaks ironic
- no reason to raise the min right now

requests >=2.25.1  ->  unchanged
- 2.27.1 has been in u-c for 1 month
- 2.26.0 drops py3.5 support
- no reason to raise the min right now

test requirements
-----------------
castellan>=3.7.0          -> >=3.10.0     in u-c for 4 months
oslo.vmware>=3.8.0        -> >=3.10.0     in u-c for 4 months
oslotest>=4.4.1           -> >=4.5.0      in u-c for 5 months
stestr>=3.1.0             -> >=3.2.1      in u-c for 4 months

coverage>=5.5  ->  unchanged
- 6.3: stopped supporting py3.6; u-c caps py3.6 at 6.2
- current is 6.3.1 which is also the u-c cap for py3.8
- no reason to raise the min right now

Change-Id: Iec14833ab502b4bb342938d5477c57742068f1b8