NVMe-oF: read mdstat in Python
The get_md_name uses the contents of mdstat using cat and piping it to
grep and then to awk using a privileged shell, but there's no need to do
that.

Actually the mdstat file has 0444 access, so we can read it in Python
directly instead of having to call the privsep daemon to create the
subprocess with the shell.

This patch changes the current implementation of the method to do the
reading and searching in Python.

Change-Id: Idbba6f9a6d928aa94d2920fc1bf8b6e2a0626a07