Support shared_targets tristate value
Cinder microversion 3.69 adds an additional value to shared_targets
beyond true and false.  Now null/None is also a valid value that can be
used to force locking.

So we now have 3 possible values:

- True ==> Lock if iSCSI initiator doesn't support manual scans
- False ==> Never lock.
- None ==> Always lock.

This patch updates the guard_connection context manager to support the
3 possible values.

With this Cinder can now force locking for NVMe-oF drivers that share
subsystems.

Closes-Bug: #1961102
Depends-On: I8cda6d9830f39e27ac700b1d8796fe0489fd7c0a
Change-Id: Id872543ce08c934cefbbbdaff6ddc61e3828b1f1